Module noise::blocks

Start here! Finished noise functions built using the primitive noise blocks.

Just create the noise type you need and start using it! If you have to make adjustments you can extend the returned values using UnboxNoise::new and then combine it with all building blocks available. (This is subject to change, once -> impl ... return types have landed in Rust.)

Take a look at the source of this module if you want to build your own noise functions.

Functions

new_noise_1d
Generates smooth (value-interpolated) one-dimensional continuous noise.

new_noise_1d_int
Generates value-interpolated one-dimensional continuous noise.

new_noise_2d
Generates two dimensional gradient noise using sensible defaults.

new_noise_2d_ex
Generates two-dimensional continuous gradient noise.

new_perlin_noise_1d
Generates coherent one-dimensional so called Perlin Noise.

new_perlin_noise_2d
Generates coherent two-dimensional Perlin Noise.

new_white_noise
Generates random (white) noise in the given bounds (both ends inclusive).