Struct noise::interpolate::CosInterpolator

pub struct CosInterpolator;

Interpolates by using the cosine slope.

This is slightly better than Perlin Interpolation but much more expensive. See Wolfram Alpha to get an idea of how this looks like.