Trait noise::interpolate::Interpolator

pub trait Interpolator<F: Float> {
    fn interpolate(&self, a: F, b: F, percent: F) -> F;
}

Interface to interpolate beweeen two numbers.

Required Methods

fn interpolate(&self, a: F, b: F, percent: F) -> F

Interpolates between the two numbers a and b. percent is a float between 0 and 1 (both borders inclusive).

There are some laws implementors of this function have to obey:

The result of interpolate(a, b, p) never changes.

interpolate(a, b, 1.0) == interpolate(b, c, 0.0)

interpolate(a, b, p) <= interpolate(a, b, q) for all p < q

The following law is sensible but not stricly required:

interpolate(a, b, p) == a + b - interpolate(a, b, 1 - p): This means that the function has to be mirrored around the center point of the interpolated area. Literally all sane interpolation functions obey this law. This also implies that interpolate(a, b, 0.5) == (a + b) / 2.