Struct noise::modules::Checkerboard

pub struct Checkerboard {
    pub size: usize,
}

Noise module that outputs a checkerboard pattern.

This noise module takes one input, size, and outputs 2size-sized blocks of alternating values. The values of these blocks alternate between -1.0 and 1.0.

This noise module is not very useful by itself, but it can be used for debugging purposes.

Fields
size: usize

Controls the size of the block in 2size.