Struct noise::modules::Select

pub struct Select<Source1, Source2, Control, T> {
    pub source1: Source1,
    pub source2: Source2,
    pub control: Control,
    pub edge_falloff: T,
    pub lower_bound: T,
    pub upper_bound: T,
}

Noise module that outputs the value selected from one of two source modules chosen by the output value from a control module.

Fields
source1: Source1

Outputs a value.

source2: Source2

Outputs a value.

control: Control

Determines the value to select. If the output value from the control module is within a range of values know as the selection range, this noise module outputs the value from source2. Otherwise, this noise module outputs the value from source1.

edge_falloff: T

Edge-falloff value.

lower_bound: T

Lower bound of the selection range.

upper_bound: T

Upper bound of the selection range.