Struct noise::Fbm

pub struct Fbm {
    pub octaves: usize,
    pub frequency: f64,
    pub lacunarity: f64,
    pub persistence: f64,
    // some fields omitted
}

Noise function that outputs fBm (fractal Brownian motion) noise.

fBm is a monofractal method. In essence, fBm has a constant fractal dimension. It is as close to statistically homogeneous and isotropic as possible. Homogeneous means “the same everywhere” and isotropic means “the same in all directions” (note that the two do not mean the same thing).

The main difference between fractal Brownian motion and regular Brownian motion is that while the increments in Brownian motion are independent, the increments in fractal Brownian motion depend on the previous increment.

fBm is the result of several noise functions of ever-increasing frequency and ever-decreasing amplitude.

fBm is commonly referred to as Perlin noise.

Fields
octaves: usize

Total number of frequency octaves to generate the noise with.

The number of octaves control the amount of detail in the noise function. Adding more octaves increases the detail, with the drawback of increasing the calculation time.

frequency: f64

The number of cycles per unit length that the noise function outputs.

lacunarity: f64

A multiplier that determines how quickly the frequency increases for each successive octave in the noise function.

The frequency of each successive octave is equal to the product of the previous octave’s frequency and the lacunarity value.

A lacunarity of 2.0 results in the frequency doubling every octave. For almost all cases, 2.0 is a good value to use.

persistence: f64

A multiplier that determines how quickly the amplitudes diminish for each successive octave in the noise function.

The amplitude of each successive octave is equal to the product of the previous octave’s amplitude and the persistence value. Increasing the persistence produces “rougher” noise.

Implementations

impl Fbm

pub const DEFAULT_SEED: u32

pub const DEFAULT_OCTAVE_COUNT: usize

pub const DEFAULT_FREQUENCY: f64

pub const DEFAULT_LACUNARITY: f64

pub const DEFAULT_PERSISTENCE: f64

pub const MAX_OCTAVES: usize

pub fn new() -> Self