Struct noise::RotatePoint

pub struct RotatePoint<Source> {
    pub source: Source,
    pub x_angle: f64,
    pub y_angle: f64,
    pub z_angle: f64,
    pub u_angle: f64,
}

Noise function that rotates the input value around the origin before returning the output value from the source function.

The get() method rotates the coordinates of the input value around the origin before returning the output value from the source function.

The coordinate system of the input value is assumed to be “right-handed” (x increases to the right, y increases upward, and z increases inward).

Fields
source: Source

Source function that outputs a value

x_angle: f64

x rotation angle applied to the input value, in degrees. The default angle is set to 0.0 degrees.

y_angle: f64

y rotation angle applied to the input value, in degrees. The default angle is set to 0.0 degrees.

z_angle: f64

z rotation angle applied to the input value, in degrees. The default angle is set to 0.0 degrees.

u_angle: f64

u rotation angle applied to the input value, in degrees. The default angle is set to 0.0 degrees.

Implementations

impl<Source> RotatePoint<Source>

pub fn new(source: Source) -> Self

pub fn set_x_angle(self, x_angle: f64) -> Self

Sets the rotation angle around the x axis to apply to the input value.

pub fn set_y_angle(self, y_angle: f64) -> Self

Sets the rotation angle around the y axis to apply to the input value.

pub fn set_z_angle(self, z_angle: f64) -> Self

Sets the rotation angle around the z axis to apply to the input value.

pub fn set_u_angle(self, u_angle: f64) -> Self

Sets the rotation angle around the u axis to apply to the input value.

pub fn set_angles(
self,
x_angle: f64,
y_angle: f64,
z_angle: f64,
u_angle: f64
) -> Self

Sets the rotation angles around all of the axes to apply to the input value.