Trait noise::Noise

pub trait Noise<In> {
    type Out;

    fn value(&self, position: In) -> Self::Out;
}

A common interface for all structs that can generate noise.

Associated Types

type Out

Required Methods

fn value(&self, position: In) -> Self::Out

Calculates the value of the noise function at the given position.

The noise for a position never changes.