Trait noise::NoiseModule

pub trait NoiseModule<T> {
    type Output;
    fn get(&self, point: T) -> Self::Output;
}

Base trait for noise modules.

A noise module is a object that calculates and outputs a value given a n-Dimensional input value, where n is (2,3,4).

Each type of noise module uses a specific method to calculate an output value. Some of these methods include:

Calculating a value using a coherent-noise function or some other mathematical function.

Mathematically changing the output value from another noise module in various ways.

Combining the output values from two noise modules in various ways.

Associated Types

type Output

Required Methods

fn get(&self, point: T) -> Self::Output