Crate noise

This crate implements functions to generate one or multi dimensional smooth and Perlin noise.

The examples folder contains several examples that show how to use this library.

Modules

blocks
Start here! Finished noise functions built using the primitive noise blocks.

interpolate
Interpolating between two numbers.

Structs

CombinedNoise
Creates noise by combining multiple source noises.

CombinedNoise2
Creates a noise that combines two source noises.

ConstantNoise
Generator that always returns the same value.

DefaultI32Noise
Generator for pseudo-random integer noise.

InputOp
Applies an operation on the position parameter before the position is passed to the source noise.

InterpolatedNoise
Takes discrete input noise to produce value-interpolated output noise.

InterpolatedNoise2D
Produces two-dimensional gradient-noise using discrete one-dimensional source noise.

NoOpNoise
Generator that always returns the position parameter unchanged.

OutputOp
Applies an operation to the output of the source noise.

UnboxNoise
Wrapper that can be used to use Box<Noise<...>> as source to other building blocks.

Traits

Noise
A common interface for all structs that can generate noise.