Trait noise::NoiseFn

pub trait NoiseFn<T> {
    fn get(&self, point: T) -> f64;
}

Base trait for noise functions.

A noise function is a object that calculates and outputs a value given a n-Dimensional input value, where n is (2,3,4).

Each type of noise function uses a specific method to calculate an output value. Some of these methods include:

Calculating a value using a coherent-noise function or some other mathematical function.

Mathematically changing the output value from another noise function in various ways.

Combining the output values from two noise functions in various ways. TODO: change to NoiseFn<T, const N: usize> & point: [T; N] when const generics are stabilized.

Required methods

fn get(&self, point: T) -> f64

Loading content...