Trait noise::Seedable

pub trait Seedable {
    fn set_seed(self, seed: u32) -> Self;

    fn seed(&self) -> u32;
}

Trait for functions that require a seed before generating their values

Required methods

fn set_seed(self, seed: u32) -> Self

Set the seed for the function implementing the Seedable trait

fn seed(&self) -> u32

Getter to retrieve the seed from the function

Loading content...